requirements.txt
flask
flask-login
werkzeug
pandas
numpy
scikit-learn
joblib
mysql-connector-python